Fix Data extension func testAsFetchToken to not use .lineLength64Characters
